Added Docker compose, caddyfile and github actions page
For a vps prod setup, having caddy is like having cert bots but with
less manual configuration /cron job.  it does come with adding another
container, because of this we need docker compose now.

Since this is a static website, we can run it directly with github
actions pages.  github will host the site for free.